Fix package application regression
The most probable cause of the regression is memory allocation.

Since SigningExtension code is not supposed to be multithreaded we can keep
a single preallocated buffer and reuse it every time. This is the fastest
approach.

If for some reason we want to make the code multithreaded with the ability
to add several entries at the same time, and we have to allocate a buffer
every time, we can reduce memory allocation delay by dynamically adjusting
the buffer size based on the entry size. The test data showed that the
regression was mostly visible on the entries of sizes <8096 bytes.
Thus, we could allocate smaller buffers for smaller entries and keep it
65536 for larger ones.

Bug: 119756507
Test: existing

PiperOrigin-RevId: 222059777
Change-Id: I3a23d6d0883d678eb5279d53db6367d6fe974d6b