Added 8-byte ZipField Support to apkzlib
This is the first CL to start adding Zip64 support into apkzlib.

Added support for 8-byte fields for the Zip64 specification. This will
lay the groundwork for the Zip64 EOCD Locator, the Zip64 EOCD, and the
Zip64 Extensible Data Sector which all require 64-bit fields. This is
written to be consistent with how the 2 and 4 byte fields are in code.

Due to a limitation in how JAVA handles longs, it is impossible to
support the high-end of unsigned 8-byte fields without using something
like BigInteger. However, since the 8-byte files are used for
location/size within the Zip File, it implies that the ZipFile is larger
than 2^63 -1 bytes (over a exabyte) of data which is well beyond what we
need to support.

PiperOrigin-RevId: 223220037
Change-Id: Ie1e249756e84307f59ef6c2f6c095b745ea82ccc