Adding Zip64 end of central directory and end of central directory locator.
This is part two of adding Zip64 support to apkzlib.

This part adds the infrastructure for the Zip64 EOCD and the locator for the EOCD. These classes are presently unconnected to the rest of the ZFile framework which will come with later changes. The next change will involve grouping these two classes with the Zip32 EOCD in order to have all of the related end of central directory record information in one place.

The tests for reading and writing the Zip64 EOCD require the use of the EOCD record and will be added in a later change.

PiperOrigin-RevId: 224220428
Change-Id: I4f07d8cb646b5fd763bbe3a100431270d0f9d972