Added EocdGroup class to handle all ZFile work related to the EOCD record.
This is part 3 of the Zip64 Support changes to apkzlib.

The ZFile class is already a massive class with a lot of verification and work logic. With the addition of the optional Zip64 EOCD Locator and Zip64 Eocd, it makes sense to move the entirety of EOCD logic into one collective spot. As such, a lot of the EocdGroup's logic is pulled directly from the ZFile class, with some slight modification to handle the cases when a Zip64 EOCD record is present. The EocdGroup class has yet to be wired into the ZFile, so the changes will not yet have an effect.

Testing has been added to test the writing and reading of the entire EocdGroup class as information stored in the EOCD dictates what information is relevant in the Zip64 EOCD.

PiperOrigin-RevId: 224925348
Change-Id: I10675ba33b206e04e94941a5a372018020103d08