Migrate org.mockito.Matchers#any* to org.mockito.ArgumentMatchers
The former is deprecated and replaced by the latter in Mockito 2. However, there is a
functional difference: ArgumentMatchers will reject `null` and check the type
if the matcher specified a type (e.g. `any(Class)` or `anyInt()`). `any()` will
remain to accept anything.

For more information see go/matchers-any-null-lsc

Tested:
    TAP --sample ran all affected tests and none failed
    http://test/OCL:250283471:BASE:250280528:1559058357636:afff21ab
PiperOrigin-RevId: 250302607
Change-Id: If6a3f08f4809b26bfa74c3416290c5f6ede8c8ff