Automated g4 rollback of changelist 225992388.
*** Reason for rollback ***

This change introduces padding in the central directory when the zip entries are aligned.
This happens because the padding between files is introduced via the extra field, which is here being copied into the corresponding entry in the central directory.

For native libraries, this padding is 4k alignment, which significantly increases the size of the central directory.

As advised in b/133494783 by kudasov, reverting also the changes in cl/226015310 and cl/227020762.

*** Original change description ***

Make extra fields in local header and central directory match

This behavior matches e.g. ZipOutputStream. Most importantly it allows to make performance
improvements since reading metadata as single block (central directory) is much cheaper than reading it separately (central directory should be read anyway).

***

PiperOrigin-RevId: 251206066
Change-Id: I161423f6a2cd2cc86372e01ca17e9d4136614a31